Add mstring-reader shorthands
Instead of the function #'use-mstrings, the mstring reader itself is
exported from the package to allow lispers to attach it to whatever
dispatch character pair they want.

This allowed the reader to pick up on subchars and change the behavior
respectively; here, if the subchar is #\" or #\>, the block mode is
assumed and the reader immediately reads the following string.