Omit #> shorthand macro on Clozure CL 1.2 and later
The #> dispatch character is claimed on CCL since 1.2, as a macro to
read a keyword preserving case. I can't complain, since it's not
reserved to the user, but none of the reserved sharpsign macros really
speak to me as "This folds the following string" -- not even #" is safe.

The #> mstring macro is elided for all systems with the :CCL-1.2
feature, and its respective test is skipped.