Fix leading whitespace on folding M-string
If a folding block mode M-string starts with an initial line, it's
replaced with a Space (as how lines are joined together), like so:

    #M>"
        Foo"

    => " Foo"

This was a surprising result, and I called it a bug. The algorithm was
adjusted to find a natural fix, and four new assertions are added to the
test suite.